moved all files to scripts and created test dir and file create .nvmrc create package.json
