+++ package.json
@@ -1,0 +1,25 @@
+{
+  "name": "mu-shamir",
+  "version": "0.0.1",
+  "description": "javascript implementation of shamir secret sharing",
+  "main": "index.js",
+  "scripts": {
+    "test": "node tests/index.js"
+  },
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/Mu-Methods/shamir.git"
+  },
+  "keywords": [
+    "shamir"
+  ],
+  "author": "Mu Methods admin@mu-methods.tech (@ReneeSauce & @frankiebee)",
+  "license": "MIT",
+  "bugs": {
+    "url": "https://github.com/Mu-Methods/shamir/issues"
+  },
+  "homepage": "https://github.com/Mu-Methods/shamir#readme",
+  "devDependencies": {
+    "tape": "^5.6.0"
+  }
+}
