+++ scripts/polynomial.js
